Added circle algorithms and bezier curve

+++ Graphics Project- V1/utils.h
@@ -4,6 +4,6 @@
 
 int Round(double x);
 struct Point {
-    double x, y;
-    Point(double x = 0, double y = 0) : x(x), y(y) {}
+    int x, y;
+    Point(int x = 0, int y = 0) : x(x), y(y) {}
 };
